fix(dash/auth): :passport_control: Check permissions on admin routes

+++ apps/dashboard/src/actions/claims.ts
@@ -1,8 +1,14 @@
 'use server';
+import { getSession, hasRole } from '@/util/auth';
 import prisma from '@/util/db';
 import { revalidatePath } from 'next/cache';
 
 export const adminChangeTeam = async (data: { claimId: string; teamId: string }) => {
+	const session = await getSession();
+	if (!hasRole(session, 'edit-claims')) {
+		throw new Error('Unauthorized');
+	}
+
 	console.log('changeTeam', data);
 	const claim = await prisma.claim.update({
 		where: {
@@ -18,6 +24,11 @@ export const adminChangeTeam = async (data: { claimId: string; teamId: string })
 };
 
 export const adminDeleteClaim = async (data: { claimId: string }) => {
+	const session = await getSession();
+	if (!hasRole(session, 'edit-claims')) {
+		throw new Error('Unauthorized');
+	}
+
 	const claim = await prisma.claim.delete({
 		where: {
 			id: data.claimId,
