+++ apps/dashboard/src/actions/contacts.ts
@@ -1,9 +1,15 @@
 'use server';
+import { getSession, hasRole } from '@/util/auth';
 import { revalidateWebsitePath } from '@/util/data';
 import prisma from '@/util/db';
 import { revalidatePath } from 'next/cache';
 
 export const adminAddContact = async (data: { name: string; role: string; email: string; discord: string }) => {
+	const session = await getSession();
+	if (!hasRole(session, 'edit-contacts')) {
+		throw new Error('Unauthorized');
+	}
+
 	const contact = await prisma.contact.create({
 		data,
 	});
@@ -20,6 +26,11 @@ export const adminEditContact = async (data: {
 	email: string;
 	discord: string;
 }) => {
+	const session = await getSession();
+	if (!hasRole(session, 'edit-contacts')) {
+		throw new Error('Unauthorized');
+	}
+
 	const contact = await prisma.contact.update({
 		where: {
 			id: data.id,
@@ -38,6 +49,11 @@ export const adminEditContact = async (data: {
 };
 
 export const adminDeleteContact = async (id: any) => {
+	const session = await getSession();
+	if (!hasRole(session, 'edit-contacts')) {
+		throw new Error('Unauthorized');
+	}
+
 	const contact = await prisma.contact.delete({
 		where: {
 			id,
